staff views helper methods moved
