+++ staff/views.py
@@ -1,10 +1,7 @@
 from django.shortcuts import render, redirect
 from django.contrib.auth.models import Group
-from home.models.models import UserProfile, Classroom
-from django.contrib.auth.models import User as DjangoUser
 from django.contrib import messages
 from home.decorators import group_required
-from home.models.salesforce import ClassEnrollment
 from home.forms import (
     CreateStaffForm,
     CreateClassroomForm,
@@ -13,6 +10,7 @@ from home.forms import (
     CreateStudentForm,
     CreateClassOfferingForm,
 )
+from .staff_views_helper import *
 
 
 @group_required("staff")
@@ -126,109 +124,43 @@ def create_volunteer_user(request):
     return render(request, "create_volunteer_user.html", {"form": form})
 
 
-def create_user_with_profile(form):
-    new_user = DjangoUser.objects.create_user(
-        username="%s.%s"
-        % (form.cleaned_data.get("first_name"), form.cleaned_data.get("last_name")),
-        email=form.cleaned_data.get("email"),
-        first_name=form.cleaned_data.get("first_name"),
-        last_name=form.cleaned_data.get("last_name"),
-        password="missionbit",
-    )
-    birthdate = str(form.cleaned_data.get("birthdate")).replace(r"/", "")
-    birthdate_year = birthdate[0:4]
-    birthdate_m = birthdate[8:]
-    birthdate_d = birthdate[5:7]
-    new_user.userprofile.change_pwd = True
-    new_user.userprofile.salesforce_id = "%s%s%s%s%s" % (
-        form.cleaned_data.get("first_name")[:3],
-        form.cleaned_data.get("last_name")[:3],
-        birthdate_year,
-        birthdate_d,
-        birthdate_m,
-    )
-    new_user.userprofile.date_of_birth = "%s-%s-%s" % (
-        birthdate_year,
-        birthdate_d,
-        birthdate_m,
-    )
-    new_user.save()
-    return new_user
-
-
 @group_required("staff")
 def create_classroom(request):
     if request.method == "POST":
         form = CreateClassroomForm(request.POST)
         if form.is_valid():
             classroom = setup_classroom_teachers(form)
-            for volunteer in form.cleaned_data.get("volunteers"):
-                enroll_in_class(form, volunteer)
-                django_user = (
-                    UserProfile.objects.filter(salesforce_id=volunteer.client_id)
-                    .first()
-                    .user_id
-                )
-                print("volid: %s" % volunteer.id)
-                classroom.volunteers.add(django_user)
-            for student in form.cleaned_data.get("students"):
-                enroll_in_class(form, student)
-                django_user = (
-                    UserProfile.objects.filter(salesforce_id=student.client_id)
-                    .first()
-                    .user_id
-                )
-                print("student id: %s" % student.id)
-                print("got user %s from salesforce: " % student)
-                classroom.students.add(django_user)
+            add_volunteers_and_students_to_classroom(form)
             messages.success(
                 request,
                 f'Classroom {form.cleaned_data.get("course")} Successfully Created',
             )
             classroom.save()
             return redirect("staff")
+        else:
+            messages.error(
+                request,
+                "Classroom NOT created, your form was not valid, please try again.",
+            )
+            return redirect(request, "staff")
     form = CreateClassroomForm()
     return render(request, "create_classroom.html", {"form": form})
-
-
-def enroll_in_class(form, user):
-    ClassEnrollment.objects.get_or_create(
-        name=form.cleaned_data.get("course"),
-        created_by=form.cleaned_data.get("created_by"),
-        contact=user,
-        status="Enrolled",
-        class_offering=form.cleaned_data.get("course"),
-    )
-
-
-def setup_classroom_teachers(form):
-    print("teach id: %s" % form.cleaned_data.get("teacher").client_id)
-    print("teach id: %s" % form.cleaned_data.get("teacher_assistant").client_id)
-    classroom = Classroom.objects.create(
-        teacher_id=UserProfile.objects.filter(
-            salesforce_id=form.cleaned_data.get("teacher").client_id
-        )
-        .first()
-        .user_id,
-        teacher_assistant_id=UserProfile.objects.filter(
-            salesforce_id=form.cleaned_data.get("teacher_assistant").client_id
-        )
-        .first()
-        .user_id,
-        course=form.cleaned_data.get("course").name,
-    )
-    enroll_in_class(form, form.cleaned_data.get("teacher"))
-    enroll_in_class(form, form.cleaned_data.get("teacher_assistant"))
-    return classroom
 
 
 @group_required("staff")
 def create_class_offering(request):
     if request.method == "POST":
         form = CreateClassOfferingForm(request.POST)
-        form.save()
-        print("FUCK YEAH!!!")
-        return redirect("staff")
+        if form.is_valid():
+            form = CreateClassOfferingForm(request.POST)
+            form.save()
+            return redirect("staff")
+        else:
+            messages.error(
+                request,
+                "Class offering NOT created, your form was not valid, please try again.",
+            )
+            return redirect(request, "staff")
     form = CreateClassOfferingForm()
     return render(request, "create_class_offering.html", {"form": form})
 
@@ -236,3 +168,5 @@ def create_class_offering(request):
 @group_required("staff")
 def my_account_staff(request):
     return render(request, "my_account_staff.html")
+
+
